+++ pyfecons/costing_data.py
@@ -194,6 +194,16 @@ class CostingData(SerializableToJSON):
             + self.cas28.C280000
         )
 
+    def overnight_cost(self) -> M_USD:
+        """CAS10 through CAS50 (before IDC)."""
+        return M_USD(
+            self.cas10.C100000
+            + self.cas20.C200000
+            + self.cas30.C300000
+            + self.cas40.C400000
+            + self.cas50.C500000
+        )
+
     def cas10_to_60_total_capital_cost(self) -> M_USD:
         return M_USD(
             self.cas10.C100000