+++ pyfecons/report/sections/cas600000_section.py
@@ -1,10 +1,7 @@
 from dataclasses import dataclass
 
 from pyfecons.costing.categories.cas600000 import CAS60
-from pyfecons.materials import Materials
 from pyfecons.report.section import ReportSection
-
-materials = Materials()
 
 
 @dataclass
@@ -13,8 +10,5 @@ class CAS60Section(ReportSection):
         super().__init__()
         self.template_file = "CAS600000.tex"
         self.replacements = {
-            "C600000": str(round(cas60.C600000)),  # TODO - not in template
-            "C610000": str(round(cas60.C610000)),
-            "C630000LSA": str(round(cas60.C630000LSA)),
-            "C630000XXX": str(round(cas60.C630000)),
+            "C600000": str(round(cas60.C600000)),
         }